traffic: working label-based traffic shifting
There are a number of caveats to the current formula. It works like so:

release A has 10 pods with a weight of 10
release B has 10 pods with a weight of 5

The total number of pods is 20. The total weight is 15.

release A gets:
    10/15 weight =
        0.66 * 20 total pods =
            13.2 -> 10 pods (only 10 pods available).

release B gets:
    5/15 weight =
        0.33 * 20 total pods =
            6.6 -> 7 pods (round up to int).

This is a bit weird: you might reasonably expect weight 10 to get
double the traffic of weight 5, not a 65/35% split, as resulted. In
other words, weights are used to create proportions. However, it is
probably reasonable enough.

Things get tricky if release B has 20 pods.

The total number of pods is 30. The total weight is 15.

release A gets:
    10/15 weight =
        0.66 * 30 total pods =
            16.66 -> 10 (only 10 pods available).

release B gets:
    5/15 weight =
        0.33 * 30 total pods =
            9.9 -> 10 pods (round up to int).

As a result, the traffic split will be 50/50 due to the difference in
how much of the total pool each release represents. This is despite the
fact that you requested that A be weighted more heavily,

This is unfortunate, but I think we can cope with it for the time being:
any service mesh will give us request-level balancing, not pod-level
balancing.