strategycontroller: WIP - Business logic sort of works
There is an issue with regards to the releaseInformer, where it doesn't
get proper updates after a release is modified. @btyler mentioned this
could be some misalignment between the Release object and the CRD I have
in my development environment, which I'll check after commit and push
this so he can take a look as well.