shipment order: fix type of replicas, default to 1 replica
Replicas isn't an optional field in the Release, so it was converted to
from *int32 to int32 in 24a9a82ae0effa63bb3b5e996f1cad052032e7c0.

As part of this change we needed to decide on a 'nil' translation when
we extract it from the Deployment object. Per
k8s.io/api/apps/v1/types.go DeploymentSpec comments, 'nil' replicas is
defaulted to 1 replica.