ci: fix up static check errors, add metalinter config
'gometalinter.v2 ./...' in the project root will run the same set of
checks as we do in CI with the same set of 'ignore' params (vendor and
the generated client code).

This commit also fixes all of the lint failures currently present in
master. Mostly they were trivial things; we can have a chat about
whether we find vetshadow useful enough to merit the noise of constantly
avoiding the `thing, err := blorg()` pattern (shadowing err happens
a lot).