strategycontroller: protect against out of bounds targetSteps
Along the way, move code out of pkg/apis/ into a utility func in
pkg/controller/strategy. If we develop more than a few of these, we can
create pkg/strategy to hold them.

This also tweaks the hardcoded strategy utility to work with full
Strategies rather than StrategySpecs, mostly so that the Name can be
used in error reporting.