cluster client store: default to use system CA, typos
The Cluster Secret controller provisions a secret with 'tls.crt' and
'tls.key' which lines up with the Kubernetes TLS type Secret. This
includes renaming 'tls.cert', a silly typo which I made early on, with
the actual key name from Kubernetes TLS secrets: tls.crt.

Note: this allows the client store to with with vanilla TLS secrets, or
with Opaque secrets that have tls.ca as well. This facilitates local and
integration testing.

You have two options for setting things up locally:

* Use the cluster secret controller and add minikube's CA to your system store. This creates a vanilla K8s TLS secret.
* Use `hack/create-minikube-cluster-secret.pl`. This creates an Opaque secret with `tls.ca` set from minikube.