traffic controller: port to cluster client store
This removes the stand-in cluster management in favor of the cluster
client store which abstracts that problem.

As a result, the API for the pod label shifter changed slightly (sync
cluster) in a way which makes the TrafficShifter interface kind of
premature.

We'll figure out Istio/$other_service_mesh when the time comes.