client store: significant revamp
This commit transforms the client store into mostly-a-controller, with
workqueues for secrets and clusters. Also, tests.

As a result, it restores the 'Run' method, which is where the store
waits for the secret/cluster informer caches to sync.

Internally a target cluster's client/informer/config is represented by
a cluster struct, access to which is serialized through a single
goroutine. This arrangement prevents replacing the existing cached
cluster unless it would change the key attributes: secret checksum or
cluster API server hostname. It also makes it easier (IMO) to reason
about order of operations when starting the new cluster's informer and
shutting down the old one.

Finally, this commit makes the client store suitable for use by multiple
controllers by changing the SubscriptionFunc and
EventHandlerRegistrationFunc into slices of callbacks with methods to
append.