cluster client store: panic -> return
This panic() was not as much of an invariant as I thought. This was due
to goroutine schedule ordering: a srv.Store() followed by srv.Stop()
could Shutdown() the cluster before WaitForInformerCache is run, which
would put it into state Terminated.

Since we only want to warm up the cache if we're still planning to use
it (StateNotReady), it is safe and reasonable to just return out of this
goroutine in the case where we've entered some other state.