controller-manager: map of init funcs like k8s
This changes the controller manager to take a map of init funcs, one per
controller it starts. This allows individual controllers to be easily
skipped, like if we start the controller manager without paths for the
Cluster Secret controller.

This also changes the signature for clusterclientstore.Run to behave
exactly like a controller.