schedule controller: hardcode the TypeMeta for OwnerRefs
When creating a new ShipmentOrder (and thus, Release and associated
Target objects) on a freshly-started controller, we observed this error:

E0308 00:54:44.780080   24546 schedule_controller.go:134] error syncing:
'default/release-v0.0.1': InstallationTarget.shipper.booking.com
"release-v0.0.1" is invalid: [metadata.ownerReferences.apiVersion:
Invalid value: "": version must not be empty,
metadata.ownerReferences.kind: Invalid value: "": kind must not be
empty]

Somehow the apiVersion and kind for a newly-made Release were empty.
This problem went away if the controller was restarted. I'm still not
sure what's going on, and why objects _sometimes_ have a TypeMeta and
_sometimes_ do not, but
https://github.com/kubernetes/client-go/issues/60#issuecomment-281747911
is a significant possibility -- in some cases the APImachinery codec is
erasing the GroupVersionKind.