Revert "Revert "Traffic controller: use TrafficTarget labels to list Services""
This reverts commit dcbeb33013b4fb3fa3d448f10a9b96fadfaf0e8b.

There aren't really any good options here: we need to cooperate with the
chart to produce a service that has the same name from release to
release, so it sticks around. We need the traffic controller to somehow
find this service. There are two approaches: pick the service by some
pre-arranged name (like $appname-prod) or by a pre-arranged label (like
shipper-lb: production).

We initially implemented the first. However, this becomes tricky: our
code doesn't have an idea of application name, only namespace. @asurikov
proposed (and I agreed) that it would be better to pick the Service by
labels; in particular using the set of labels from the ShipmentOrder
releaseSelectors, which basically represent application identity. This
was sourced from the TrafficTarget labels, which contain Release among
them.

The problem with this was that including 'Release' as a label on
a fixed-name service which lasts for multiple releases is at best
confusing, at worst impossible to implement label-based load balancing
for.

As a result, we reverted that change back to fetching the service by
name.

However, after some more thinking, I beieve it's slightly less bad to
have an API that consists of an extra label on a service in the
application chart rather than a specially-named object. It's still
a contract with the chart, but it's a flexible one that does not lock us
into namespace == app.