dep ensure update: k8s/apiserver is gone
When I consolidated the executables into cmd/shipper, I removed the
dependency on k8s.io/apiserver, which we were using for
SetupSignalHandler. I'm comfortable with the function we have in
cmd/shipper, so I think it makes sense to have our own implementation
and be able to flush this fairly significant set of deps.