app: working draft with app.Status.History instead of pointers
This removes the release queue from the App controller in favor of
making any release update enqueue the associated Application object.

This draft works and passes a basic test, but has some known bugs:

* no mutexing on rollouts, so creating a new release while the previous
is deploying will just march onwards
* the strategy controller uses listers in callback handlers, which is
unreliable (especially in tests)
* Replicas and Clusters are still part of the Release Environment, and
therefore need to be excluded from hashing so that the app controller
does not re-create the same release over and over as they get scheduled.
They should move to annotations or a field outside of 'environment'