app: add basic sanity check for hashReleaseEnvironment
This function is used all over as a helper, and the errors we get from
our tests when it misbehaves are a bit cryptic. This is kind of testing
the stdlib, but hashReleaseEnvironment has a little bit of extra sauce
(json marshal) that's a candidate to clean up down the line.

On balance I think this helps more than it hurts, even if it only helps
a little.