CI/CD: the "CD" part
This introduces three new stages to our GitLab pipelines: 'build',
'release', and 'deploy'. They only trigger when a tag is pushed.

- build: create a go binary out of cmd/shipper/main.go
- release: use that binary to create a docker image, tagged with $CI_COMMIT_TAG
- deploy: patch a K8s Deployment object in the management cluster to reference the fresh image

These stages are automatic and contingent on all jobs of the previous
stage succeeding.

As a result, any new tag which passes testing and linting will update
the version of shipper running in the management cluster.