chart cache: fix flapping cache eviction test
This is an interesting case. I had a bug in my code where I was reverse
sorting instead of regular sorting.

However, the tests passed 999/1000 times on OSX because the mtime
resolution on HFS+ is 1 second, and the tests almost always completed
inside of a single second. As a result, the StableSort meant that the
order given by ioutil.ReadDir was preserved: this order is alphabetical
by filename: this coincided with the order that the tests expected.

However, the GitLab pipeline runner has a filesystem with higher
resolution mtime. In consequence, the threshold for test duration to
trigger the failing test was much lower, and the test would flap
approximately 1/4 runs.

This commit reverses the sort order (fixing the bug). I explored making the test
case use an order distinct from alphabetical so this bug is exposed more
easily, but then we have the opposite problem: the test fails on OSX
almost all the time.