traffic controller: do not include 'foreign' pods in weight calculations
The Pod label traffic shifter was fetching the total pod count for the
entire namespace, not for a single Application. As a result, an
Application with two releases (A and B) each with weight 100 and two
pods, would actually achieve a weight of 80 in a namespace which had
a fifth, unassociated pod.

The math goes like this for a single release:

0.4 (2 pod / 5 total pods) * 200 total weight (100 + 100) = 80

The bug is that there are 5 pods, not 4. This patch corrects it to read
as follows:

0.5 (2 pods / __4__ total pods) * 200 total weight (100 + 100) = 100.

As a side effect, this patch changes the interface for
newPodLabelShifter to use 'appName' as the resource identifier instead
of the collection of labels applied to the TrafficTarget object.