e2e tests: fix flapping test due to incorrect pod check
The tests were counting 'Ready' pods based on the pod.Status. However,
the ReplicaSet controller in K8s filters out any pod with
a DeletionTimestamp. As a result we were counting Pods that looked
perfectly 'Ready', but which had been marked for deletion.

This meant that our tests thought incumbents had too many pods, when all
the controllers reported them scaled down to 0.