schedule controller: introduce basic region and capability scheduling
This patch makes the schedule controller translate a ClusterRequirements
struct into a set of concrete clusters.

This means: every required region must be represented, and every region
must have all the capabilities listed as a requirement.

The scheduler currently makes no statement about required level of
redundancy: how many clusters in each region. The application will be
added to every cluster in the region which meets the requirements.