charts: exclude whitespace-only objects from manifests to install
It's tricky to use Go templating in a Helm chart to make sure you will
produce exactly zero bytes: consider '{{ end -}}' vs '{{ end }}'.

This makes Shipper a bit more tolerant: anything with whitespace-only
contents will be skipped when building up a list of manifests.

This is useful when you have a chart which conditionally excludes
a particular object, like a staging Service which you don't want in
development.