schedule controller: introduce 'Replicas' to ClusterRequirements
This patch introduces the ability for users to request a particular
number of clusters in a given region, like so:

```yaml
clusterRequirements:
  regions:
  - name: us-east
    replicas: 3
  capabilities:
  - gpu
```

This means that the Schedule controller will schedule this Release on
3 clusters in us-east, all with the 'gpu' capability. If there are not
enough clusters in that region, or in that region with that capability,
scheduling will fail.

Along the way, this patch introduces weighted rendezvous hashing[^1] in
order to 'stick' Applications to clusters in a way which is maximally
resilient to clusters coming or going, and provides some basic load
balancing.[^2]

How it works
------------

This works by creating a "preference list" for each Application: a list
of clusters sorted in a way unique to this Application.  If one of those
clusters is gone, unschedulable, or substituted, the Application will
shift minimally: likely the Application will change by only one cluster.

Changes to Cluster objects
--------------------------

This change includes two new fields to Cluster objects, Seed and Weight.

**Weight** is a multiplier against a cluster's natural location in the
preference list. Higher weights raise the cluster's position. This can
be used to gradually phase-in or phase-out a cluster.

**Seed** is the cluster portion of the sorting key for the preference list
(the other portion is Application specific). This enables cluster
operators to 'impersonate' a given cluster: for example, if we want to
shift all the workload in cluster A specifically to cluster B without
redistributing cluster A's workload across the fleet.

The chosen hashing function for the weighted rendezvous hashing is
xxHash[^3], which is fast and has good distribution properties.

[^1]: See [wikipedia](https://en.wikipedia.org/wiki/Rendezvous_hashing),
[a nice blog post](https://www.pvk.ca/Blog/2017/09/24/rendezvous-hashing-my-baseline-consistent-distribution-method/),
and
[slides from a talk introducing the idea of weights](http://www.snia.org/sites/default/files/SDC15_presentations/dist_sys/Jason_Resch_New_Consistent_Hashings_Rev.pdf)

[^2]: Hash-based load balancing converges on random load distribution,
which allows for large differences in load between clusters. Once we
start incorporating cluster capacity information into scheduling we'll
be doing something similar to [consistent hashing with bounded
loads](https://ai.googleblog.com/2017/04/consistent-hashing-with-bounded-loads.html):
as we traverse the preference list we'll skip over clusters which are
too heavily loaded.

[^3]: http://cyan4973.github.io/xxHash/