types: stop lying about AchievedStep 0; add 'name'
A bugfix and a feature:

Bugfix:

release.Status.AchievedStep was a value type, not a pointer type. As
a result, when a Release is created, it already shows 'achievedStep 0',
even though that is not the case: installation, capacity, or traffic may
have failed.

Feature:

User feedback indicated it'd be nice to include the 'name' of the
strategy step in the achievedStep status.