testing: framework for using clusterclientstore in unit tests
This introduces testing.ClusterClientStore, which generates the required
clusters, secrets, clients, and informers to be able to use the real
ClusterClientStore in unit tests.

The unit under test is a 'cluster' fixture, which has actions and
objects like any other fixture.

It also adds CheckClusterClientActions, which uses the actions and
objects from all the cluster fixtures to check testing outcomes.