traffic controller: filter out excessive statuses
The Traffic controller works with all TrafficTarget objects for a given
app at once: it needs to do this because 'weight' is relative.

As a result, it calculates a cluster status for every cluster it
encounters while syncing, regardless of whether that cluster is
specified by the currently-syncing TrafficTarget object.

This patch removes the calculated statuses for 'foreign' clusters from
the syncing TrafficTarget's .Status.Clusters field. This should allow the
Strategy controller to do cluster-migration rollouts: previously it got
stuck because it has a check forbidding unknown clusters in status
fields.