all: v1 -> v1alpha1
When I created the types initially, I was unfamiliar with the
significance of picking a particular API version. As a result, I picked
'v1'. This is not where Shipper is at the moment: we firmly expect some
of our APIs to grow and evolve, potentially drastically, as it matures.

One editorial decision: 'shipperv1' was renamed to 'shipper', not
'shipperv1alpha1'. This is consistent with K8s controllers as far as
I can tell.