all: raise kubeclient QPS settings
We can see from our metrics that we're pegged at the default QPS limit
(~5 QPS) on more populated clusters. However, we don't know where our
true ceiling is -- and as a result, we do not know the current desired
ratio of API calls to live Shipper objects, which is a primary
efficiency metric.

This commit substantially raises QPS settings so that we can:

1) stop giving clients a bad experience, and
2) get information on that key efficiency metric.

"Substantially" here means 10x: each controller previously had a 5 QPS
limit; now it has 50. The shared client for the remote clusters raised
from 5 (a bug, it was meant to be 15) to 150. In aggregate this takes us
from ~30qps limit with ~60qps burst to ~400qps limit, with ~800qps
burst.

As long as we have capacity to spare (and we believe we do: we're
reasonably certain that this ~400-800qps ballpark is within API
server+etcd capability based on casual benchmarking indicating capacity
in the 4-6k QPS range) we'd like to raise our sensitivity to efficiency
improvements by raising the limit.