Implemented e2e tests for aborts refs #11
This commit extends the end-to-end test suite with 2 new test cases:
one ofr a new application and the other one for an existing one.
In both cases, we remove the current active release object and watch
what's happening. In the case of a new application we expect the system
to revert everything to step 0 and wait for a command, whereas an
existing application, which is being rolled forward, should pick up the
previous existing release and move full-on to the last step of this
release.