SchedulerController treats already-exists as errors #11
This commit changes the behavior of schedulercontroller.scheduler in
such a way where it treats already-exists cases (installation target,
capacity target and traffic target) as errors.

The reason for this change is the recent investigation on stall tests
from #11. What was discovered is: a release deletion triggers kubernetes
to garbage-collect all release children objects. This happens
asynchronously, and sometimes a new application installation finds these
orphan objects and could not distinguish them from the normal ones, so
it immediately returns. This behavior is being changed by this very
commit: if we detect that an object exists but we were planning to
create a new one, we return an error immediately.