installer: comment & error message tweak
It isn't a `panic()` anymore, and probably it should not be, since it
boils down to whether the Release has `shipper-app` as a label.

The TODO and error message are pointers for a broader story about making
sure that one can use Release objects independently of the Application
controller setup.