shipperctl: add support for external auth providers
This allows it to work with, for example, GKE clusters with the 'gcp'
auth provider in ~/.kube/config.

Also tweak the quick install docs to mention that you need to be a real
cluster-admin. On GKE you need to create a clusterrolebinding first;
otherwise you see errors that indicate trying to create the new Shipper
clusterroles is privilege escalation.

This commit includes a giant pile of new stuff in vendor: this is
everything all of the cloud providers for their auth plugins.