Split release scheduling in 2 independent stages
This commit splits release controller handler in 2 independent stages:
cluster allocation and associated object creation.
The first step chooses clusters and assigns them on the release object.
If this operation succeeds, we return from the handler right away and
rely on the notification mechanism to call the handler once the
replication chain is done propagating the first change. On the second
round we create associated objects for the release and enqueue a new
object in the application queue. The rest got no modifications since the
last implementeation milestone.