clusterclientstore: take resync period as a parameter
We are currently investigating if Shipper works correctly when resyncs
are disabled for #77. Before we can carry on with that investigation,
though, we need to ensure that Shipper does in fact disable all resyncs.
As far as the logs show, this is the only piece that currently doesn't
honor cmd/shipper -resync.