installation controller: check contender only once
For reasons I couldn't exactly ascertain, we were trying to check if the
InstallationTarget's Release is the current contender release: once by
asking the lister, and once by checking the Application's history.

The first one works alright, but the second one suffers from a race
condition: the application controller would create a release, hand off
work to the release controller to create the installation target, and
only then update the Application resource's history. The second check in
the installation controller would then fail if it executed before the
Application was updated, since the history would be out of date. This
has only worked before because of resyncs, and in the context of testing
shipper with resyncs disabled (see #77) we eventually found this race
condition.

As we have this check twice, this was easy: let's just remove the broken
one :)