Made Makefile customizable and moved `webhook` out of the `controllers` directory
I also removed the `install` make target because I didn't find a way
of replacing `bookingcom/shipper:latest` with the specified
environment variable reliably. If people want to install Shipper on
their cluster, they can always modify the file and do `kubectl apply
-f kubernetes/deployment.shipper.yaml`.

I left `webhook` in the list of controllers in `cmd/shipper/main.go`
because it already provides a lot of boilerplate to allow users to
specify if they want to have a webhook running or not. I can't think
of a way to have the ability of enabling/disabling the webhook on the command
line without introducing a lot of repeated code.

Apart from that, I also added the check to `webhook.go` to check if
the error returned from `listenAndServe` was caused by a server close,
or if it was an actual error.