driveby: fix for with range clause and pointers
I found it very odd when seeing this over and over in our logs:

```
 Found 2 pod conditions with the status set to `false`. The first has a
 type of ContainersReady, and the second has a type of ContainersReady.
```

It didn't sound right that we'd have two `ContainersReady` conditions.
It turns out that we don't, in fact we have one `ContainersReady` and
another `Ready` condition, but since a for statement with a range clause
reuses the `condition` variable declared in the for statement [1], we
were getting an address pointing to memory being changed over and over.

Now, we create a copy of the struct, and get a pointer to *that*, so Go
can't sweep values from under us anymore.

[1] https://golang.org/ref/spec#For_range