janitor controller: rename files
This makes logs slightly easier to read, since glog reports on the name
of the file, not the name of the package. Having someone report itself
as "controller.go" doesn't exactly tell us which controller it is, so
renaming the janitor controller files solves that, while also falling in
line with our other controllers.