application controller: handle errors from cleanUpReleasesForApplication
This addresses another part of #77, trying to make shipper work with
resyncs disabled.

Before this change, cleaning up the list of releases was "fire and
forget": we wouldn't check if any errors happened, and would just keep
on trucking as if nothing happened. This is fine when you can expect
resyncs to pick up the slack, but since we're specifically trying to
avoid that, we need to check whenever errors happen, and retry
appropriately.