shipper-state-metrics: collect time spent on pending release conditions
We're trying to collect metrics on how long rollouts take, end to end.
Although we can't currently get that directly, we have a decent proxy
metric: how long a release has spent with a Strategy.Condition.Status=False.

We will start exposing that through a prometheus summary metric through
shipper-state-metrics, with the 50th, 90th and 99th percentile. I chose
not to go with a histogram because we currently have no idea about the
distribution of these numbers, so it's really hard to set up buckets
that make sense. Also, I don't expect many aggregations to be done on
this, as we only have one instance of shipper-state-metrics running,
collecting data for all available releases.

For getting these numbers back, for instance the 90th percentile we've
been waiting for releases with ContenderAchievedInstallation=False, one
could query prometheus with the following:

	shipper_release_durations{cond_type="ContenderAchievedInstallation", quantile="0.9"}