More defensive CapacityTarget spec population
This commit is a followup on a problem with uncontrollable
CapacityTarget Status Clusters inflation. The new fleet summary
functionality introduced an extended reporting in CapacityTarget.Status,
which might case a problem: `etcdserver: mvcc: database space exceeded`.

This change limits the amount of items in Status.Clusters explicitly and
makes sure the old reports are being cleaned up.