Introduce structured errors (#91)
* errors: introduce structured errors

* capacity controller: use structured errors

* driveby: fix for with range clause and pointers

I found it very odd when seeing this over and over in our logs:

```
 Found 2 pod conditions with the status set to `false`. The first has a
 type of ContainersReady, and the second has a type of ContainersReady.
```

It didn't sound right that we'd have two `ContainersReady` conditions.
It turns out that we don't, in fact we have one `ContainersReady` and
another `Ready` condition, but since a for statement with a range clause
reuses the `condition` variable declared in the for statement [1], we
were getting an address pointing to memory being changed over and over.

Now, we create a copy of the struct, and get a pointer to *that*, so Go
can't sweep values from under us anymore.

[1] https://golang.org/ref/spec#For_range

* janitor controller: rename files

This makes logs slightly easier to read, since glog reports on the name
of the file, not the name of the package. Having someone report itself
as "controller.go" doesn't exactly tell us which controller it is, so
renaming the janitor controller files solves that, while also falling in
line with our other controllers.

* janitor controller: use structured errors

* janitor controller: don't fire and forget

* clusterclientstore: use structured errors

* release controller: use structured errors

* controller: move errors to pkg errors

* driveby: remove duplicated func

getAssociatedApplicationName already existed in a slightly different
form in pkg releaseutils as ApplicationNameForRelease. It didn't return
a more specialized error type, but this commit addresses that as well.

* errors: better interface for KubeclientError

* application controller: use structured errors

* installation controller: use structured errors

* installation controller: rename files

This just renames controller_test.go to installation_controller_test.go,
to match the code file installation_controller.go. This also makes
vim-go happier when using :GoAlternate to navigate between code and test
file ^_^

* clustersecret controller: use structured errors

* traffic controller: use structured errors

* errors: introduce MultiError.Flatten()

* errors: improve documentation